+++ autd3-core/Cargo.toml
@@ -47,7 +47,8 @@ gain = ["defined", "datagram", "geometry", "thiserror", "bit-vec", "zerocopy", "
 geometry = ["nalgebra", "bvh", "paste", "getset", "defined", "derive-new", "derive_more", "derive_more/add", "derive_more/mul", "derive_more/into_iterator", "derive_more/deref", "derive_more/debug"]
 left_handed = []
 link = ["zerocopy", "getset", "ethercat", "datagram", "geometry", "derive_more", "derive_more/display"]
-modulation = ["getset", "utils", "defined", "datagram", "derive_more", "derive_more/display"]
+modulation = ["getset", "sampling_config", "utils", "defined", "datagram", "derive_more", "derive_more/display"]
+sampling_config = []
 use_meter = []
 utils = ["windows"]
