+++ autd3-modulation-audio-file/src/error.rs
@@ -1,6 +1,6 @@
 use std::num::ParseIntError;
 
-use autd3_core::{derive::ModulationError, modulation::SamplingConfigError};
+use autd3_core::{derive::ModulationError, sampling_config::SamplingConfigError};
 use thiserror::Error;
 
 #[derive(Error, Debug)]
